+++ umap/static/umap/js/umap.controls.js
@@ -562,107 +562,6 @@ L.Control.Embed = L.Control.Button.extend({
   },
 })
 
-U.DataLayer.include({
-  renderLegend: function (container) {
-    if (this.layer.renderLegend) return this.layer.renderLegend(container)
-    const color = L.DomUtil.create('span', 'datalayer-color', container)
-    color.style.backgroundColor = this.getColor()
-  },
-
-  renderToolbox: function (container) {
-    const toggle = L.DomUtil.createButtonIcon(
-      container,
-      'icon-eye',
-      L._('Show/hide layer')
-    )
-    const zoomTo = L.DomUtil.createButtonIcon(
-      container,
-      'icon-zoom',
-      L._('Zoom to layer extent')
-    )
-    const edit = L.DomUtil.createButtonIcon(
-      container,
-      'icon-edit show-on-edit',
-      L._('Edit')
-    )
-    const table = L.DomUtil.createButtonIcon(
-      container,
-      'icon-table show-on-edit',
-      L._('Edit properties in a table')
-    )
-    const remove = L.DomUtil.createButtonIcon(
-      container,
-      'icon-delete show-on-edit',
-      L._('Delete layer')
-    )
-    if (this.isReadOnly()) {
-      L.DomUtil.addClass(container, 'readonly')
-    } else {
-      L.DomEvent.on(edit, 'click', this.edit, this)
-      L.DomEvent.on(table, 'click', this.tableEdit, this)
-      L.DomEvent.on(
-        remove,
-        'click',
-        function () {
-          if (!this.isVisible()) return
-          if (!confirm(L._('Are you sure you want to delete this layer?'))) return
-          this._delete()
-        },
-        this
-      )
-    }
-    L.DomEvent.on(toggle, 'click', this.toggle, this)
-    L.DomEvent.on(zoomTo, 'click', this.zoomTo, this)
-    container.classList.add(this.getHidableClass())
-    container.classList.toggle('off', !this.isVisible())
-  },
-
-  getHidableElements: function () {
-    return document.querySelectorAll(`.${this.getHidableClass()}`)
-  },
-
-  getHidableClass: function () {
-    return `show_with_datalayer_${L.stamp(this)}`
-  },
-
-  propagateDelete: function () {
-    const els = this.getHidableElements()
-    for (const el of els) {
-      L.DomUtil.remove(el)
-    }
-  },
-
-  propagateRemote: function () {
-    const els = this.getHidableElements()
-    for (const el of els) {
-      el.classList.toggle('remotelayer', this.isRemoteLayer())
-    }
-  },
-
-  propagateHide: function () {
-    const els = this.getHidableElements()
-    for (let i = 0; i < els.length; i++) {
-      L.DomUtil.addClass(els[i], 'off')
-    }
-  },
-
-  propagateShow: function () {
-    this.onceLoaded(function () {
-      const els = this.getHidableElements()
-      for (let i = 0; i < els.length; i++) {
-        L.DomUtil.removeClass(els[i], 'off')
-      }
-    }, this)
-  },
-})
-
-U.DataLayer.addInitHook(function () {
-  this.on('hide', this.propagateHide)
-  this.on('show', this.propagateShow)
-  this.on('erase', this.propagateDelete)
-  if (this.isVisible()) this.propagateShow()
-})
-
 const ControlsMixin = {
   HIDDABLE_CONTROLS: [
     'zoom',
