+++ umap/templates/umap/js.html
@@ -5,6 +5,9 @@
 <script type="module"
         src="{% static 'umap/js/modules/leaflet-configure.js' %}"
         defer></script>
+<script src="{% static 'umap/vendors/markercluster/leaflet.markercluster.js' %}"
+        defer></script>
+<script src="{% static 'umap/vendors/heat/leaflet-heat.js' %}" defer></script>
 {% if locale %}
   {% with "umap/locale/"|add:locale|add:".js" as path %}
     <script src="{% static path %}" defer></script>
@@ -21,14 +24,11 @@
 <script src="{% static 'umap/vendors/csv2geojson/csv2geojson.js' %}" defer></script>
 <script src="{% static 'umap/vendors/osmtogeojson/osmtogeojson.js' %}" defer></script>
 <script src="{% static 'umap/vendors/loading/Control.Loading.js' %}" defer></script>
-<script src="{% static 'umap/vendors/markercluster/leaflet.markercluster.js' %}"
-        defer></script>
 <script src="{% static 'umap/vendors/contextmenu/leaflet.contextmenu.min.js' %}"
         defer></script>
 <script src="{% static 'umap/vendors/photon/leaflet.photon.js' %}" defer></script>
 <script src="{% static 'umap/vendors/georsstogeojson/GeoRSSToGeoJSON.js' %}"
         defer></script>
-<script src="{% static 'umap/vendors/heat/leaflet-heat.js' %}" defer></script>
 <script src="{% static 'umap/vendors/fullscreen/Leaflet.fullscreen.min.js' %}"
         defer></script>
 <script src="{% static 'umap/vendors/toolbar/leaflet.toolbar.js' %}" defer></script>
@@ -49,7 +49,6 @@
 <script src="{% static 'umap/js/umap.features.js' %}" defer></script>
 <script src="{% static 'umap/js/umap.permissions.js' %}" defer></script>
 <script src="{% static 'umap/js/umap.datalayer.permissions.js' %}" defer></script>
-<script src="{% static 'umap/js/umap.layer.js' %}" defer></script>
 <script src="{% static 'umap/js/umap.controls.js' %}" defer></script>
 <script src="{% static 'umap/js/umap.js' %}" defer></script>
 <script src="{% static 'umap/js/components/fragment.js' %}" defer></script>
